Move `event` into history/ because name conflicts
