+++ src/game/status-menu.js
@@ -7,7 +7,7 @@ import * as UI from 'style';
 import ColorPicker from 'color-picker';
 
 import * as Game from 'game';
-import * as Event from 'event';
+import * as Event from 'history/event';
 import * as server from 'server';
 import * as Stream from '../stream';
 import { SetConnectionCtx } from 'connection';
