+++ src/record/otherEvents.js
@@ -5,7 +5,7 @@ import * as humanTime from 'humanTime';
 import * as UI from 'style';
 
 import * as Game from 'game';
-import * as Event from 'event';
+import * as Event from 'history/event';
 
 type PlayerJoinProps = {
     +event: Event.PlayerJoin,
